Fixup compile-time hash design, primarily adapting to 64-bit values
1) Default hash generations are now 64-bit
2) Default method now FNV-1a
3) Horner hash available, and original CRC32 one can be used directly

Switcher for defaults/inclusion needs work, but is simply unclean for now

+++ src/core/services/config/ConfigServer.h
@@ -43,9 +43,9 @@ struct cvar
 	std::string  attrib; ///< attribute name
 	std::string  value;  ///< data value
 	std::string  default_value;  ///< data value default
-	uint32_t     hash;   ///< hash value for this setting
+	size_t       hash;   ///< hash value for this setting
 
-	cvar(const char* vpath, const char* vattrib, const char* vvalue, const char* vdvalue, uint32_t vhash)
+	cvar(const char* vpath, const char* vattrib, const char* vvalue, const char* vdvalue, size_t vhash)
 		: path(vpath), attrib(vattrib), value(vvalue), default_value(vdvalue), hash(vhash)
 	{
 	}
